# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OmniSight RSM",
    "description": "Ресурсно-сервисная модель (РСМ) — API v1 для учёта элементов, связей и корреляции алертов.\n\n**Integration API (ТЗ §8)** — `Resources`, `Correlation`, `Meta`: контракт для correlation engine и внешних систем мониторинга.\n\n**Admin API** — `CI`, `Relations`, справочники, Autodiscover, аудит: ведение модели через UI и администрирование.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/resources/resolve": {
      "post": {
        "tags": [
          "Integration API · Resources (§8)"
        ],
        "summary": "Resolve V1",
        "operationId": "resolve_v1_api_v1_resources_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequestV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/resources/search": {
      "get": {
        "tags": [
          "Integration API · Resources (§8)"
        ],
        "summary": "Search Resources V1",
        "operationId": "search_resources_v1_api_v1_resources_search_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "hostname",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hostname"
            }
          },
          {
            "name": "ip",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ip"
            }
          },
          {
            "name": "externalId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Externalid"
            }
          },
          {
            "name": "serviceCode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Servicecode"
            }
          },
          {
            "name": "applicationCode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Applicationcode"
            }
          },
          {
            "name": "cmdbId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cmdbid"
            }
          },
          {
            "name": "type_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type Name"
            }
          },
          {
            "name": "owner",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/{resource_id}/relations": {
      "get": {
        "tags": [
          "Integration API · Resources (§8)"
        ],
        "summary": "Resource Relations V1",
        "description": "§8: прямые связи объекта РСМ.",
        "operationId": "resource_relations_v1_api_v1_resources__resource_id__relations_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/{resource_id}": {
      "get": {
        "tags": [
          "Integration API · Resources (§8)"
        ],
        "summary": "Resource Detail V1",
        "description": "§8: карточка объекта РСМ.",
        "operationId": "resource_detail_v1_api_v1_resources__resource_id__get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/{resource_id}/graph": {
      "get": {
        "tags": [
          "Integration API · Resources (§8)"
        ],
        "summary": "Resource Graph V1",
        "operationId": "resource_graph_v1_api_v1_resources__resource_id__graph_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resource Id"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10,
              "minimum": 1,
              "default": 2,
              "title": "Depth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/{resource_id}/impact": {
      "get": {
        "tags": [
          "Integration API · Resources (§8)"
        ],
        "summary": "Resource Impact V1",
        "operationId": "resource_impact_v1_api_v1_resources__resource_id__impact_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/{resource_id}/components": {
      "get": {
        "tags": [
          "Integration API · Resources (§8)"
        ],
        "summary": "Resource Components V1",
        "operationId": "resource_components_v1_api_v1_resources__resource_id__components_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/{resource_id}/business-path": {
      "get": {
        "tags": [
          "Integration API · Resources (§8)"
        ],
        "summary": "Business Path V1",
        "operationId": "business_path_v1_api_v1_resources__resource_id__business_path_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/{resource_id}/graph-layout": {
      "get": {
        "tags": [
          "Integration API · Resources (§8)"
        ],
        "summary": "Get Graph Layout V1",
        "operationId": "get_graph_layout_v1_api_v1_resources__resource_id__graph_layout_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resource Id"
            }
          },
          {
            "name": "relation_filter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Relation Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Integration API · Resources (§8)"
        ],
        "summary": "Save Graph Layout V1",
        "operationId": "save_graph_layout_v1_api_v1_resources__resource_id__graph_layout_put",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resource Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphLayoutUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Integration API · Resources (§8)"
        ],
        "summary": "Clear Graph Layout V1",
        "operationId": "clear_graph_layout_v1_api_v1_resources__resource_id__graph_layout_delete",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resource Id"
            }
          },
          {
            "name": "relation_filter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Relation Filter"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlertIdentifier": {
        "properties": {
          "hostname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hostname"
          },
          "ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip"
          },
          "serviceCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Servicecode"
          },
          "applicationCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicationcode"
          },
          "externalId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Externalid"
          },
          "cmdbId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cmdbid"
          }
        },
        "type": "object",
        "title": "AlertIdentifier"
      },
...
      "GraphLayoutPosition": {
        "properties": {
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          }
        },
        "type": "object",
        "required": [
          "x",
          "y"
        ],
        "title": "GraphLayoutPosition"
      },
      "GraphLayoutUpdate": {
        "properties": {
          "relation_filter": {
            "type": "string",
            "title": "Relation Filter",
            "default": "*"
          },
          "positions": {
            "additionalProperties": {
              "$ref": "#/components/schemas/GraphLayoutPosition"
            },
            "type": "object",
            "title": "Positions"
          }
        },
        "type": "object",
        "title": "GraphLayoutUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResolveRequestV1": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Page Size",
            "default": 100
          },
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/AlertIdentifier"
            },
            "type": "array",
            "title": "Alerts"
          }
        },
        "type": "object",
        "required": [
          "alerts"
        ],
        "title": "ResolveRequestV1"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT из `POST /api/v1/auth/token` (поле `session.access_token`). В Authorize вставьте только токен."
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Webhook-Secret",
        "description": "Секрет webhook-интеграций."
      },
      "XApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Ключ интеграции для correlation ingest и внешних систем."
      }
    }
  },
...
}